Cleanup of webpack config :)
